home and allClient tabs working, i think
